Rebase to HEAD as of a few days ago.

+++ sys/arch/alpha/include/cdefs.h
@@ -1,8 +1,8 @@
-/* $NetBSD: cdefs.h,v 1.12 2012/01/20 14:08:05 joerg Exp $ */
+/* $NetBSD: cdefs.h,v 1.12.6.1 2014/08/20 00:02:41 tls Exp $ */
 
 #ifndef _MACHINE_CDEFS_H_
 #define	_MACHINE_CDEFS_H_
 
-#define	__ALIGNBYTES		7
+#define	__ALIGNBYTES		((size_t)7)
 
 #endif /* !_MACHINE_CDEFS_H_ */
